Renamed data folder to resources and added sample data
