+++ test/resources/sample.json
@@ -1,0 +1,6 @@
+{
+  "key1": "value1",
+  "value2": {
+    "nested": "pair"
+  }
+}
